Added SafariBooksOnline.com as Read resources
Safari Books Online LLC is a digital library. Founded as a joint venture between O'Reilly Media and the Pearson Technology Group.

For a relatively small fee, compared to the amount of content available, numerous tech books can be read online. This resource constantly updates it's content store and offers content well before it is officially published.